requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-docx>=0.8.11
openpyxl>=3.1.2
streamlit-authenticator>=0.2.3
PyYAML>=6.0
bcrypt>=4.0.0